requirements.txt
streamlit==1.22.0
pandas==1.5.3
numpy==1.24.3
python-binance==1.0.17
plotly==5.14.1
python-dotenv==1.0.0
requests==2.28.2
pytz==2023.3
google-cloud-bigquery==3.9.0
discord.py==2.2.2
aiohttp==3.8.4